Split transit_hub.rs (754 lines) into subdirectory modules
Break the monolithic transit_hub.rs into three focused files:
- types.rs (268 lines): constants, TransitMode, TransitHubType, TransitHub
  component, TransitHubs registry, TransitHubStats, and Saveable impls
- systems.rs (210 lines): hub detection, land value, stats systems + plugin
- tests.rs (268 lines): all unit tests

All pub items re-exported from mod.rs so external imports unchanged.

Part of #1389
